+++ base-compat-batteries/src/Data/Foldable1/Compat/Repl/Batteries.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE PackageImports #-}
+{-# OPTIONS_GHC -fno-warn-dodgy-exports -fno-warn-unused-imports #-}
+-- | Reexports "Data.Foldable1.Compat"
+-- from a globally unique namespace.
+module Data.Foldable1.Compat.Repl.Batteries (
+  module Data.Foldable1.Compat
+) where
+import "this" Data.Foldable1.Compat